Move entities and setting double doctrine connexions configuration
