+++ .env.dist
@@ -14,6 +14,7 @@ APP_SECRET=73a6b907a6b5d96e6bbe3804e186708f
 # For an SQLite database, use: "sqlite:///%kernel.project_dir%/var/data.db"
 # Configure your db driver and server_version in config/packages/doctrine.yaml
 DATABASE_URL=mysql://db_user:db_password@127.0.0.1:3306/db_name
+CUSTOMER_DATABASE_URL=mysql://db_user:db_password@127.0.0.1:3306/customer_db_name
 ###< doctrine/doctrine-bundle ###
 
 ###> nelmio/cors-bundle ###
