+++ config/packages/doctrine.yaml
@@ -7,29 +7,46 @@ parameters:
 
 doctrine:
     dbal:
-        # configure these for your database server
-        driver: 'pdo_pgsql'
-        server_version: '5.7'
-        charset: utf8
-        default_table_options:
-            charset: utf8mb4
-            collate: utf8mb4_unicode_ci
-
-        url: '%env(resolve:DATABASE_URL)%'
+        connections:
+            default:
+                driver:   pdo_pgsql
+                charset:  UTF8
+                url:      '%env(resolve:DATABASE_URL)%'
+                server_version: '5.7'
+                default_table_options:
+                    charset: utf8mb4
+                    collate: utf8mb4_unicode_ci
+            customer:
+                driver:   pdo_pgsql
+                charset:  UTF8
+                url:      '%env(resolve:CUSTOMER_DATABASE_URL)%'
+                server_version: '5.7'
+                default_table_options:
+                    charset: utf8mb4
+                    collate: utf8mb4_unicode_ci
     orm:
+        default_entity_manager: default
         auto_generate_proxy_classes: '%kernel.debug%'
-        naming_strategy: doctrine.orm.naming_strategy.underscore
-        auto_mapping: true
-        mappings:
-            PiaApi:
-                is_bundle: false
-                type: annotation
-                dir: '%kernel.project_dir%/src/Entity'
-                prefix: 'PiaApi\Entity'
-                alias: PiaApi
-            PiaOauth:
-                is_bundle: false
-                type: annotation
-                dir: '%kernel.project_dir%/src/Entity/Oauth'
-                prefix: 'PiaApi\Entity\Oauth'
-                alias: PiaOauth
+        entity_managers:
+            default:
+                auto_mapping: true
+                naming_strategy: doctrine.orm.naming_strategy.underscore
+                connection: default
+                mappings:
+                    PiaOauth:
+                        is_bundle: false
+                        type: annotation
+                        dir: '%kernel.project_dir%/src/Entity/Oauth'
+                        prefix: 'PiaApi\Entity\Oauth'
+                        alias: PiaOauth
+            customer:
+                auto_mapping: false
+                naming_strategy: doctrine.orm.naming_strategy.underscore
+                connection: customer
+                mappings:
+                    PiaApi:
+                        is_bundle: false
+                        type: annotation
+                        dir: '%kernel.project_dir%/src/Entity/Pia'
+                        prefix: 'PiaApi\Entity\Pia'
+                        alias: PiaApi
