+++ src/Controller/CommentController.php
@@ -13,7 +13,7 @@ namespace PiaApi\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use FOS\RestBundle\View\View;
 use FOS\RestBundle\Controller\Annotations as FOSRest;
-use PiaApi\Entity\Comment;
+use PiaApi\Entity\Pia\Comment;
 
 class CommentController extends PiaSubController
 {
